refactored dot benchmark
